compose: try performing recipient completion on send (fixes #1495)
This commit performs completion on recipient fields when the send button
is clicked (uncompleted text is usually present if the cursor is on the
recipient field at that time).

If any completion was performed, sending is quietly aborted. This avoids
sending mail to the wrong recipient if the uncompleted text doesn't
resolve to what the user thought.