K9-31 User Agent added on decrypt message
When you are decrypting, you got all the original meessage information
wich includes the user agent. We replace the encrypted user agent
header with the original one, so we solve the bug.

The original idea to solve this bug was using the optional fields you
have in the pEpMessage at MimeMessageBuilder, but optional fields
are not being passed correctly :(