K9-40 When Key filename length too long - no import
When getting the header parameter of the attached file, we were 
receiving a null, so we try to encode the content with a null charset.
Now, if we receive a null we use utf-8 as default charset, so we fix two things:

first: no nullpointer exception when attaching a file with really long filename
on mobile.

second: now, we can display the full attached file name instead
"noname"