Base Activities now extend AppCompatActivity
In K9Activity was not big deal, even it's implemented the code to use a toolbar, in in, in K9ListActivity is a bit different. We've had to implement ListActivity methods to still giving support to the activities that uses this. For this has been also necessary to create a list_activity layout copied from Google's codebase because we cannot acces to internal resources.