Navigation View header functionalities implemented
First of all, sorry for this commit. In it there is lot of happy programming, but I wanted to make sure I commit something functional before refactoring it (also, it's been a day without comminting nothing and I'm a bit afraid.

Basically, I have implemented all the functionalities expected in the header: switching accounts and opening menus .