Messages obtained in an Action
Messages has been moved to some kind of "domain layer", wich is responsible of obtaining messages from a cursor in a worker thread.
Cursor mapping is done inside the Action and not in a mapper because the Action is tightly coupled to the Android framework and I don't want to propagate this coupling to more layers.