P4A-533 no message needed for calculating rating when is outgoing message
Previous logic was: when you go back from trustword, get the rating from the
localmessage. this is not acceptable for outgoing messages, because
in this very case, there is no local message nor message referecence. For
the simply reason the message hasn't been created yet.

So, on activity result we must use the outgoing message rating caluculus.