P4A-431 Saved state of the toolbar
Before a fragment is destroyed, it writes its current state to a Parcel, to be used in restoring its state when recreated. The Fragment's saved state will include the state of its child Views, and since you've nested a Toolbar within it, it attempts to save and restore that, as well. The Toolbar class does not have an inner class SavedState necessary for that, so this process fails when trying to restore the Toolbar instance from the Parcel.

The solution is put the toolbar in another ViewGroup; for example, a vertical LinearLayout, or a RelativeLayout.