P4A-558 Blindly fixed recyclerview crash
By telling itself its own direction.

Because the real crash is not in search, is this: Attempt to read from field 'android.support.v7.widget.ViewInfoStore android.support.v7.widget.RecyclerView.mViewInfoStore' on a null object reference

and this issue usually shows up when there are problems with the layout manager.