P4A-644 New trustwords design
All the feature has been moved into PEpStatus.

We use ExpandableRecyclerView to show the trustwords expanded when
the button is clicked. For that, we need to disable each row's listener
and rely on the button.

It's been necessary to modify PEpIdentity's model in order to reuse it
with this expandable recyclerview.