P4A-740: Disable sleep service call
This is a workaround to try if it really helps or not.

FIXME Workaround for P4A-740: This should avoid to keep wake look while
waiting before trying again.
Needs to be tested to see the real impact i may improve the battery usage.
Need to value if does it worth to rewrite it as a JOB or not.