P4A-920 Destroy all loaders in onRestart() if any account was deleted.
The problem: When we start MessageList as NewInstance after deleting current account, the lifecycle is onRestart -> onStart -> MessageListFragment.onStart -> MessageList.onNewIntent -> Create new MessageListFragment -> MessageListonResume -> MessageListFragment.onCreate -> onStart -> onResume...

On MessageListFragment.onStart, the LoaderManagerImpl (which extends MutableLiveData) receives the lifecycle change and is setActive. This triggers LoaderManagerImpl.onChanged and starts the loader(s) (with a query to the deleted account).
Before this has any chance to happen, we destroy the loaders in onRestart() in MessageList, which happens when the user is back from SettingsActivity.

This fixes the problem whether we have only one accountuuid in the fragment (we are checking a normal folder), or several accounts (unified inbox or allmessages).