P4A-947 Destroy LocalMessageLoader on pause MessageViewFragment.
Because of the LiveData present within the LoaderManager, this loader is being called twice. This was causing the dialog to be displayed twice. Since now onPause the loader is destroyed, it cannot possibly be started in onStart by accident, and we will start it in onResume with method displayMessage() as always.