P4A-934 Convert InsertableHtmlContent from Serializable to Parcelable.
The transaction overflow is happening due to a serialized object due to
that we decided to move from Serializable implementation to the Parcelable
where we have more control about it.

caveat: We also save/restore the Strings builders, but doing some tests
seems not mandatory and already covered, but as we are not aware 100% of
where they are using we decided to save/restore.