P4A-547-P4A-923 Prevent db files to be recreated after Account was deleted.
This is the issue happening when user goes back and forth between AccountSetupBasicsFragment and AccountSetupNames. Basically Account is fully synced every time and because of Account syncing, etc, Commands are added to the command queue in MessagingController. Just then we go back to AccountSetupBasicsFragment and Account gets deleted, and so we have some Commands with outdated Account references in the queue. Some of these Commands will try to get the database file for the Account after it was deleted.