P4A-1031 Add callback for getOpenFolder in MlfUtils.
Use callback in Fragment, implementing a static class in order to prevent memory leaks.

Initialize loaders in the callback. This is done so that the heavy operation buildWhereClause that is performed in onCreateLoader happens when we already displayed the ui at least.

Extract method callMlfUtilsGetOpenFolderWithCallback as the only point to call MlfUtils.getOpenFolderWithCallback.

This is only for the case of "singleFolderMode". For special accounts case we do not call MlfUtils.getOpenFolder, so callback is not called and instead we use a GlobalLayoutListener to call initializeLoaders, so that the heavy operation in onCreateLoader happens after ui is displaying.

Show ProgressBar in onResume and hide it in onLoadFinished. This is done so that we will only hide the ProgressBar when we are really ready to show the messages (MessageListAdapter was changed).