P4A-1031 Simplify callbacks as lambdas.
Remove unneeded enumeration class "GetOpenFolderCase".
Since the callback reference in MlfUtils will not live longer than the getOpenFolder method, it is safe and less verbose to do it in this way. Nonetheless, we need to keep checking whether the Fragment is resumed before executing our callback. This just adds a minimum of lifecycle-awareness to the callback.