P4A-1051 Check for Account null in onSaveInstanceState.
The problem is that Fragment's onSaveInstanceState is called when same method of Activity is called. This means that if the Fragment was replaced (it is not visible anymore): On orientation changes we still get onSaveInstanceState, onDestroy, onCreate callbacks. And since we were initializing our Fragment's sate in onViewCreated, this means that when our Fragment is not visible anymore Account instance is not initialized. We just do not save it in that case.

Also change buttonWasVisible to wasLoading. This name makes more sense, also default is false which is less prone to mistakes.