P4A-1239 Fix MessagingControllerTest.
Mock static method PEpProviderFactory.createAndSetupProvider so that it can provide a mock of PEpProvicer. (otherwise we are trying to use the real one and get an error because we are using JNI).

Change argument matcher any(Class) in favor of any() when we have cases of calls with null values. (Since any(Class) does not match nulls).

Remove warnings from MessagingControllerTest.