P4A-1071 Remove MessageViewFragment on showing list.
We force removing MessageViewFragment in MessageList.showMessageList(). The reason is that when we go back from a message to the list and change device's orientation at the same time, MessageList.mViewSwitcher is a different instance, and so its listener's method MessageList.onSwitchComplete is not called, and MessageViewFragment not removed. This is causing that we always have a MessageViewFragment added to our activity.