P4A-1256 Find fragment by id.
In this way we find the current active fragment.

Also change the check in onActivityResult from fragment.isVisible() to fragment.isAdded().
It seems that because of the setting "Dont't keep activities", fragment's lifecycle is changed, for example fragment's onCreate is called before activity's onCreate.
Because of these lifecycle changes in the fragment, fragment.isVisible() is not yet true at the time.