P4A-1266 Update NUMBER_OF_DEVICE_NOTIFICATIONS.
We added the new OFFSET_APP_DIDNT_ENCRYPT_MESSAGE_NOTIFICATION, this is, a new notification id, that means we need to allot one more slot for notification ids for each account, so that one account notification ids do not overlap those for another account.

In this way, when we want to get a notification id for the account with getBaseNotificationId, we will get a base integer that is ahead enough of the ids for account with previous account number.