P4A-1321 First approach to sending messages before deleting account.
Create method RemoveAccountPresenter.onRemoveAccountConfirmedByUser() which checks current state. if we have any messages pending to send, then we try to send them. We will do this by default as well.

Once messages are sent, we check again if we still have any messages present in Outbox folder, in order to know whether sending messages was successful or not.

Finally, if sending messages failed we present dialog to the user again, but this time accept button action will only remove account.