P4A-1321 Extract method in MessagingController for sending messages.
Extract method MessagingController.sendPendingMessagesAndHandleSendingNotificationSynchronous, so that we can also show and hide sending messages notification while we are at the task of sending pending messages.

We choose to extract and call this synchronous method instead of relying on listener callbacks, since those could be triggered from other thread so using a listener would not reliable in our case.
Also we make our result more exact since we do not rely in any return, instead we check again whether there are still any messages sitting on the Outbox folder.

Add a delay so that our progress dialog is displayed for at least 600 ms.