P4A-1336 Remove and recreate shortcut on account deletion
This is done so that we do not get any inconsistency, like having our compose shortcut set to a null account, or having a default account but we do not have a shortcut. We remove shortcut if we are deleting the default account, and if there is still any available account afterwards we create it again.