P4A-1341 When we have no default account set, open first account
When we have no default account set in Preferences, this means there are no available accounts at all. In that case we open first account in Preferences.getAccounts. This allows us to open an account that is not enabled, and so we will prompt the user for a password to activate that account.