P4A-1073 Fix Privacy Status text and badge cut-off.
Make text autosizeable.
Use the width of the longest line in the text to measure the text, and when this is measured, center the text in PEpPrivacyStatusLayout.

Do not use wrap_content for any dimension of the MeasuredWidthTextView. Instead use only definite numbers.

Use different max autosize text size depending on device screen width.