P4A-1447 Decouple operation text updating and reading.
Done in order to prevent race conditions, we want operation text to be updated in same thread that ActivityListener.synchronizeMailboxFinished and ActivityListener.synchronizeMailboxFailed since they are setting mAccount variable to null. In this way, when we check for null in mAccount within getActionInProgressOperation, it will be a consistent and reliable result.