P4A-1101 Use LayoutStateMonitor in SplashActivity.
The first time we run the app (after install or update), if screen width is big enough, app will choose by default split view mode "always". If we are using a foldable device and device is unfolded when installing/updating app, the app will be deceived to think we are on a tablet. So by using LayoutStateMonitor we can tell apart the situation of a foldable device in unfolded mode.
In the case of a foldable device, we use same default value for split view mode as in a normal device that has at least 600dp as its greatest width.