P4A-1101 Remove "always" option also on foldable device.
If we are running on a foldable, unfolded device, the app will think it is running on a tablet and show "always" option as well as a consequence. But we can check now for unfolded foldable device. So we remove "always" option when device is not a tablet (is not be big enough to show split view correctly all the time).