PEPA-35 Set account description
Set account description after certificate is added for outgoing settings.

The fact that we arrive to this point means server settings check failed due to an untrusted certificate. After that certificate is added to keystore to be trusted, set account description to continue with setup process since server settings were fixed.