P4A-1179 Change statement order
First delete account in MessagingController. We do it before deleting account from Preferences, so that we can unset sync flag on account identity. Also we do it before deleting account from LocalStore, so that we can minimize situations of MessagingController trying to access to the local store of a deleted account.

Then delete account in Preferences, we do it early so that we can easily check if an account is deleted.

Finally we delete account in Preferences again, in order to delete possible recreated instances of remote or local stores.