P4A-1460 Show event description
When event invite has an available event description and message content does not contain that event description, replace the message content with that event description.

We also check if message content is or not longer than the event description, and choose the description only when we assume it has more information than message content.