PEPA-208 Fix MessageCompose not finding draft
Use message after moving to Drafts folder to get the MessageReference to create a new AppDidntEncryptMessageException for the notification. Reason is that on moving message to different folder, message.uid changes to a new, random one, so we wouldn't be able to get this message from database with a wrong uid.

Keep original stacktrace.