E3-10 Refractored code
Engine is now gotten as a ThreadLocal, as such the get is overriden and by default an instance is generated whenever the engine is required, should the thread hold none.
This means less boiler plate (createEngineInstanceIfNeeded()) is required before every engine call, just engine.get() is enough to always get an engine instance.