EFA-105 Update stop event time to that of late message
When we are logging an audit logger stop event, check if last log timestamp is later than the stop event timestamp. And in positive case update the stop event timestamp to that of the last log.
This provides better accuracy for stop event timestamp in this corner case. This situation can happen:
    - when a message was sent/received after last stop log was written, and before app was killed.
    - when the app is killed just after startup. This could be due to user, system or a crash.