EFA-348 Try to get last handhsake information updated
While PlanckSyncRepository is running in the background it should be able to update the last handshake init signal at any moment.
Additionally, if there is an update while the sync dialog is awaiting the user, we take it. And only when the user wants to actually display the handshake screen we lock last handshake in.