EFA-150 Reset provisioning account settings before updating
Remove provisioning settings for accounts that are not installed or being installed so the settings for those accounts will be empty or wrong if it does not come alright from the server. Then the app understands that these accounts cannot be setup and a warning will be displayed (if not displayed yet).