EFA-334 Do not mark decrypted messages as read
We just keep messages in place as they were before but decrypted, so the user can operate with them normally from the beginning.
In case of moving a Dangerous message to Suspicious folder we remove the new notification to avoid the user clicking on it and trying to open the message in the old location.